LCSD-3615 - Refactor services to use 'UseHttpClient' rather than singleton / transient in services startup.
