+++ cllc-interfaces/BCEP/BCEP.csproj
@@ -5,6 +5,7 @@
 	<ProjectGuid>{05301D75-ACF9-45D8-B572-AC052EAD4B54}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.0.0" />
     <PackageReference Include="System.ServiceModel.Duplex" Version="4.6.0" />
     <PackageReference Include="System.ServiceModel.Http" Version="4.6.0" />
     <PackageReference Include="System.ServiceModel.NetTcp" Version="4.6.0" />
