+++ cllc-interfaces/PDF/IPdfService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Gov.Lclb.Cllb.Interfaces
+{
+    public interface IPdfService
+    {
+        Task<byte[]> GetPdf(Dictionary<string, string> parameters, string template);
+    }
+}
