+++ cllc-interfaces/PDF/PDF.csproj
@@ -6,6 +6,7 @@
 	
   </PropertyGroup>
  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.0.0" />
     <PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.2.2" />
     <PackageReference Include="Microsoft.Rest.ClientRuntime" Version="2.3.20" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
